Move new files to their correct folders
