+++ lib/transposeMimeTypes.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const mimeTypeMap = require('./mime-types.json');
+const mimeTypeMap = require('../data/mime-types.json');
 
 module.exports = function transpose(map) {
   if (!map) {
